requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
bcrypt==3.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
confusable-homoglyphs==3.2.0
cryptography==37.0.4
debugpy==1.6.6
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Limiter==3.1.0
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.2.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
mock==4.0.3
numpy==1.23.2
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pandas==1.4.3
pandas-alchemy==0.0.2
password-validator==1.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.4.0
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.2
rich==12.6.0
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
username-validator==0.0.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
